src/main/java/com/ober/springmongorecipeapp/converters/CategoryCommandToCategory.java
package com.ober.springmongorecipeapp.converters;

import com.ober.springmongorecipeapp.commands.CategoryCommand;
import com.ober.springmongorecipeapp.domain.Category;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

@Component
public class CategoryCommandToCategory implements Converter<CategoryCommand, Category> {

    @Synchronized
    @Nullable
    @Override
    public Category convert(CategoryCommand source) {
        if (source == null)
            return null;

        final Category category = new Category();
        category.setId(source.getId());
        category.setDescription(source.getDescription());
        return category;
    }
}
